making the browser tests pass (had to drop some. I'll add them back on v2)
